Add files via upload
Hi everyone, 👋 

I created a dashboard to visualize data from a customer satisfaction survey. 📈

 The dashboard makes it easy to see 👀 trends in customer satisfaction over time, identify areas where there are opportunities for improvement, and track the impact of changes that have been made. 🎯


The project involved the following steps:

* 🧼 Data cleaning: I cleaned the data to remove any errors or inconsistencies.

* ⚙️ Data processing: I processed the data to create meaningful metrics and insights.

* 🔍 Data analysis: I analyzed the data to identify trends and patterns.

* 🎨 Dashboard creation: I created a dashboard using Excel to visualize the data and make it more actionable.

The dashboard was a success! 🎉 It helped the company to identify areas where there were opportunities for improvement, and it has been used to track the impact of changes that have been made.


If you're interested in learning more about this project, I'd be happy to share more details. You can also find the dashboard on my LinkedIn profile. 🔗 Sidhant Ahuja


#Excel
#businessanalysis
#powerbi 
#datavisualization 
#Dashboard 
#DataAnalysis 
#CustomerSatisfaction